+++ src/main.ts
@@ -1,20 +1,49 @@
 import PluginError from "plugin-error";
-import Config from "./config";
-import { RawConfig } from "./raw-config";
+import { RawConfig, Bundle } from "./raw-config";
 import { PluginName } from "./plugin-details";
 import { Transform, TransformCallback } from "stream";
 import * as Vinyl from "vinyl";
+import { resolve as resolvePath } from "path";
+import Catchment from "catchment";
 
 // Foward public exports
 export { MergeRawConfigs, ValidateRawConfig } from "./raw-config";
 
+// TODO How do vinyl files get renamed?
+
 export default class Bundler extends Transform {
-    private Config: Config;
 
     /**
      * Tracks all files resolved within virtual directory tree.
      */
-    private VirtualDirTree: Map<string, Vinyl>;
+    private ResolvedFiles: Map<string, (Vinyl & VinylExtension)> = new Map();
+
+    /**
+     * Used in conversion of canonical paths to virtual paths (for scenarios with resource overriding, etc).
+     * First string is an absolute path, second is a matching virtual path.
+     */
+    private PathTransforms: [string, string][] = [];
+
+    /**
+     * Script bundles to build.
+     */
+    private ScriptBundles: Map<string, string[]> = new Map();
+
+    /**
+     * Style bundles to build.
+     */
+    private StyleBundles: Map<string, string[]> = new Map();
+
+    /**
+     * 
+     */
+    private Joiner: Joiner;
+
+    /**
+     * Map containing output filenames for each bundle.
+     * Key is bundle name, value is the file path.
+     */
+    public ResultsMap: Map<string, string> = new Map();
 
     /**
      * 
@@ -23,16 +52,46 @@ export default class Bundler extends Transform {
      */
     constructor(config: RawConfig, joiner: Joiner) {
         super();
-        try {
-            // Process configuration
-            this.Config = new Config(config);
 
-            // Create virtual directory
-            this.VirtualDirTree = new Map();
+        // Extract path transformations (if set) and make canonical paths absolute
+        if (config.PathTransforms) {
+            for (const [path, vPath] of config.PathTransforms)
+                this.PathTransforms.push([resolvePath(path), vPath]);
         }
-        catch (exception) {
-            throw new PluginError(PluginName, exception);
+
+        // Add bundles
+        if (config.bundle) {
+            for (const name in config.bundle) {
+                if (config.hasOwnProperty(name)) {
+                    const bundle = config.bundle[name];
+                    // JS
+                    if (bundle.scripts)
+                    this.ScriptBundles.set(name, bundle.scripts);
+
+                    // CSS
+                    if (bundle.styles)
+                        this.StyleBundles.set(name, bundle.styles);
+                }
+            }
         }
+
+        this.Joiner = joiner;
+    }
+
+    /**
+     * Attempts to create a virutal path from the provided path.
+     * On failure, the provided path is returned.
+     * @param path Absolute path to try and resolve.
+     */
+    private ResolveVirtualPath(path: string): [string, number] {
+        // Try to resolve a virtual path
+        for (const [index, [pathStart, vPathStart]] of this.PathTransforms.entries()) {
+            if (path.startsWith(pathStart))
+                return [path.replace(path, vPathStart), index];
+        }
+
+        // No matches
+        return [path, 0];
     }
 
     /**
@@ -45,16 +104,19 @@ export default class Bundler extends Transform {
     _transform(chunk: any, encoding: string, callback: TransformCallback): void {
         if (Vinyl.isVinyl(chunk)) {
             // Grab virtual path
-            const virtualPath = this.Config.ResolveVirtualPath(chunk.path);
+            const [virtualPath, precedence] = this.ResolveVirtualPath(chunk.path);
 
-            // See if anything lines up with it in the VirtualDirTree and handle accordingly
-            if (this.VirtualDirTree.has(virtualPath)) {
-                // Let configuration resolve collision
-                this.VirtualDirTree.set(
-                    virtualPath,
-                    this.Config.ResolvePreferedFile(this.VirtualDirTree.get(virtualPath), chunk));
+            // Extended chunk
+            let file: (Vinyl & VinylExtension) = chunk as (Vinyl & VinylExtension);
+            file.path = virtualPath;
+            file.Precedence = precedence;
+
+            // Add to resolved files (handling collisions according to precedence)
+            if (this.ResolvedFiles.has(file.path)) {
+                if (this.ResolvedFiles.get(file.path).Precedence < file.Precedence)
+                    this.ResolvedFiles.set(file.path, file);
             }
-            else this.VirtualDirTree.set(virtualPath, chunk);
+            else this.ResolvedFiles.set(file.path, file);
         }
         else {
             // Push incompatible chunk on through
@@ -70,9 +132,33 @@ export default class Bundler extends Transform {
      * @param callback 
      */
     _flush(callback: TransformCallback): void {
-        // Make sure all bundle requirements are satisfied
+        // Track names of generated bundles (for integrations map)
 
         // Create bundles
+
+        // NOTE Use "catchment" to grab bundle stream results
+        // Script bundles
+        for (const paths of this.ScriptBundles) {
+            // Get files
+
+            // Create joiner instance
+
+            // Push in data and catch result
+
+        }
+
+        // Style bundles
+        for (const paths of this.StyleBundles) {
+            // Get files
+
+            // Create joiner instance
+
+            // Push in data and catch result
+
+        }
+
+        // Indicate flush is complete
+        callback();
     }
 }
 
@@ -91,4 +177,16 @@ export interface Joiner {
      * @param name Name of bundle.
      */
     Styles(name: string): Transform;
+}
+
+/**
+ * An extension interface used for providing type hinting to modified Vinyl files.
+ * This is intended for internal use only, so extensions should be removed once they are no longer needed.
+ */
+interface VinylExtension {
+    /**
+     * Represents the Vinyl instances priority.
+     * Used to override files correctly.
+     */
+    Precedence: number
 }