+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 This release focuses on simplifying the package for UserFrosting 4 to improve maintainability. Features unsupported by UF4 are largely removed.
 
+### Added
+- Raw configuration validation.
+- Raw configuration merging.
+
 ### Changed
 - Results file contains plain file paths by default (former `plain` bundle option).
 - Reads stream chunks (`Vinyl` instances) instead of directly from file system.